# docs/api.yaml
openapi: 3.0.3
info:
  title: FinBank — Mini Core Banking System API
  version: 1.0.0
  description: |
    REST API for the Mini Core Banking System.
    Built with FastAPI + MongoDB. Uses JWT Bearer authentication.
  contact:
    name: Team 2
    email: team2@fintech.edu.tr
servers:
  - url: http://localhost:8000
    description: Local Development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserRegister:
      type: object
      required: [email, password, full_name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        full_name:
          type: string
        role:
          type: string
          enum: [admin, customer]
          default: customer
    UserLogin:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
        role:
          type: string
...
    AccountCreate:
      type: object
      required: [account_type, currency]
      properties:
        account_type:
          type: string
          enum: [checking, savings, business]
        currency:
          type: string
          default: TRY
...
paths:
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegister"
      responses:
        "201":
          description: User registered
        "409":
          description: Email already exists
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login and get JWT token
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/UserLogin"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid credentials
...
  /api/v1/accounts/:
    post:
      tags: [Accounts]
      summary: Open a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountCreate"
      responses:
        "201":
          description: Account opened
    get:
      tags: [Accounts]
      summary: List my accounts
      responses:
        "200":
          description: List of accounts
  /api/v1/accounts/{account_id}/balance:
    get:
      tags: [Accounts]
      summary: Get account balance (computed from ledger)
      parameters:
        - name: account_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Balance response
...